Update Function Names to Better Reflect Intended Scope
