+++ rfsocinterface/core/data/__init__.py
@@ -21,7 +21,7 @@ from rfsocinterface.core.data.routines import (
     RemoveElectronicsNoise,
     RoutineInputs,
     RoutineResult,
-    compute_templates,
+    _compute_templates,
     decode_tone_indices,
     register_routine,
 )
@@ -82,7 +82,7 @@ __all__ = [
     'RemoveElectronicsNoise',
     'RoutineInputs',
     'RoutineResult',
-    'compute_templates',
+    '_compute_templates',
     'decode_tone_indices',
     'generate_calibrated_data',
     'get_channel_group_name',
